Move price adjust to system
