+++ src/components/trade.ts
@@ -1,5 +1,25 @@
 import { Commodity } from "../economy/commodity";
+import { Faction } from "../economy/faction";
 import { perCommodity } from "../utils/perCommodity";
+import { Budget } from "./budget";
+
+export interface Transaction extends TradeOffer {
+  commodity: Commodity;
+  time: number;
+}
+
+export interface TransactionInput extends TradeOffer {
+  commodity: Commodity;
+  faction: Faction;
+  budget: Budget;
+  allocations: Record<
+    "buyer" | "seller",
+    {
+      budget: number | null;
+      storage: number | null;
+    }
+  >;
+}
 
 export type TradeOfferType = "buy" | "sell";
 
@@ -15,6 +35,11 @@ export const startingPrice = 100;
 
 export class Trade {
   offers: TradeOffers;
+  lastPriceAdjust = {
+    time: 0,
+    commodities: perCommodity(() => 0),
+  };
+  transactions: Transaction[] = [];
 
   constructor() {
     this.offers = perCommodity(
